feat(image-picker): introduce new image picker component with improved UI
