--- apps/web/src/app/[lang]/[username]/_components/profile-banner/index.tsx
@@ -1,1 +1,0 @@
-export * from './profile-banner'
