+++ apps/web/src/components/image-picker/image-picker.tsx
@@ -1,0 +1,5 @@
+import { ImagePickerRoot } from './image-picker-root'
+
+export const ImagePicker = {
+  Root: ImagePickerRoot,
+}
