+++ apps/web/src/components/image-picker/index.ts
@@ -1,0 +1,1 @@
+export * from './image-picker'
