+++ apps/web/src/components/videos/videos.tsx
@@ -1,4 +1,4 @@
-import { Video, tmdb } from '@plotwist/tmdb'
+import { type Video, tmdb } from '@plotwist/tmdb'
 
 export type VideosProps = {
   tmdbId: number
